Move current account info to shared component
And the kicker, start using it on the sign transaction page.
