+++ ui/components/TopMenu/TopMenuProfileButton.tsx
@@ -1,16 +1,13 @@
 import React, { ReactElement } from "react"
+import SharedCurrentAccountInformation from "../Shared/SharedCurrentAccountInformation"
 
 export default function TopMenuProfileButton(props: {
-  address: string
-  nickname?: string
-  avatar?: string
   onClick?: () => void
 }): ReactElement {
-  const { address, nickname, avatar, onClick } = props
+  const { onClick } = props
   return (
     <button type="button" onClick={onClick}>
-      {nickname ?? address}
-      <div className="avatar" />
+      <SharedCurrentAccountInformation />
       <style jsx>
         {`
           button {
@@ -19,15 +16,6 @@ export default function TopMenuProfileButton(props: {
             display: flex;
             align-items: center;
             user-select: none;
-          }
-          .avatar {
-            border-radius: 12px;
-            width: 32px;
-            height: 32px;
-            background-color: white;
-            margin-left: 8px;
-            background: url("${avatar ?? "./images/portrait.png"}");
-            background-size: cover;
           }
         `}
       </style>
